sceIoLseek() should sign extend error codes.
< 0 means error, and 0x0000000080000000 is not < 0.

+++ Core/HLE/sceIo.cpp
@@ -573,7 +573,7 @@ s64 sceIoLseek(int id, s64 offset, int whence) {
 		return (u32) f->asyncResult;
 	} else {
 		ERROR_LOG(HLE, "sceIoLseek(%d, %i, %i) - ERROR: invalid file", id, (int) offset, whence);
-		return error;
+		return (s32) error;
 	}
 }
